+++ src/PdfLexer/Fonts/Glyph.cs
@@ -52,14 +52,14 @@ namespace PdfLexer.Fonts
 
     public readonly struct UnappliedGlyph
     {
-        public UnappliedGlyph(Glyph? glyph, float shift, int bytes=0)
+        public UnappliedGlyph(Glyph? glyph, decimal shift, int bytes=0)
         {
             Glyph = glyph;
             Shift = shift;
             Bytes = bytes;
         }
         public readonly Glyph? Glyph;
-        public readonly float Shift;
+        public readonly decimal Shift;
         public readonly int Bytes;
     }
 }
